# Speed up query results copying