--- a/extensions/azuremonitor/package.json
+++ b/extensions/azuremonitor/package.json
@@ -209,7 +209,7 @@
     "update-grammar": "node ../../build/npm/update-grammar.js Microsoft/vscode-azuremonitor ./syntaxes/azuremonitor.tmLanguage"
   },
   "dependencies": {
-    "dataprotocol-client": "github:Microsoft/sqlops-dataprotocolclient#2.0.0",
+    "dataprotocol-client": "github:Microsoft/sqlops-dataprotocolclient#2.0.1",
     "figures": "^2.0.0",
     "find-remove": "1.2.1",
     "@microsoft/ads-service-downloader": "^1.2.1",