--- a/extensions/datavirtualization/package.json
+++ b/extensions/datavirtualization/package.json
@@ -107,7 +107,7 @@
   "dependencies": {
     "@microsoft/ads-extension-telemetry": "^3.0.1",
     "@microsoft/ads-service-downloader": "^1.2.1",
-    "dataprotocol-client": "github:Microsoft/sqlops-dataprotocolclient#2.0.0",
+    "dataprotocol-client": "github:Microsoft/sqlops-dataprotocolclient#2.0.1",
     "vscode-nls": "^5.2.0"
   },
   "devDependencies": {