--- a/src/sql/azdata.proposed.d.ts
+++ b/src/sql/azdata.proposed.d.ts
@@ -892,10 +892,6 @@ declare module 'azdata' {
 		 * The selected ranges to be copied.
 		 */
 		selections: SelectionRange[];
-		/**
-		 * Whether to copy results from the UI process.
-		 */
-		copyFromUIProcess?: boolean;
 	}
 
 	export interface CopyResultsRequestResult {