--- a/src/sql/workbench/api/common/extHostDataProtocol.ts
+++ b/src/sql/workbench/api/common/extHostDataProtocol.ts
@@ -420,7 +420,7 @@ export class ExtHostDataProtocol extends ExtHostDataProtocolShape {
 		return this._resolveProvider<azdata.QueryProvider>(handle).saveResults(requestParams);
 	}
 
-	override $copyResults(handle: number, requestParams: azdata.CopyResultsRequestParams): Thenable<azdata.CopyResultsRequestResult> {
+	override $copyResults(handle: number, requestParams: azdata.CopyResultsRequestParams): Thenable<void> {
 		const provider = this._resolveProvider<azdata.QueryProvider>(handle);
 		if (provider.copyResults) {
 			return provider.copyResults(requestParams);