--- a/src/sql/workbench/services/query/common/queryRunner.ts
+++ b/src/sql/workbench/services/query/common/queryRunner.ts
@@ -32,6 +32,7 @@ import { ITextResourcePropertiesService } from 'vs/editor/common/services/textRe
 import { ICapabilitiesService } from 'sql/platform/capabilities/common/capabilitiesService';
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { ServerConnID } from 'sql/workbench/services/query/common/query';
+import { getRemoteAuthority, getRemoteName } from 'vs/platform/remote/common/remoteHosts';
 
 /*
 * Query Runner class which handles running a query, reports the results to the content manager,
@@ -495,11 +496,19 @@ export default class QueryRunner extends Disposable {
 	 * @param includeHeaders [Optional]: Should column headers be included in the copy selection
 	 */
 	async copyResults(selections: Slick.Range[], batchId: number, resultId: number, includeHeaders?: boolean): Promise<azdata.CopyResultsRequestResult> {
+		let copyInBackend = false;
+		const remoteAuthority = getRemoteAuthority(URI.parse(this.uri));
+		const remoteName = getRemoteName(remoteAuthority);
+		if (!remoteName) {
+			copyInBackend = true;
+		}
+
 		return this.queryManagementService.copyResults({
 			ownerUri: this.uri,
 			batchIndex: batchId,
 			resultSetIndex: resultId,
 			includeHeaders: includeHeaders,
+			copyInBackend: copyInBackend,
 			selections: selections.map(selection => {
 				return {
 					fromRow: selection.fromRow,
@@ -618,10 +627,12 @@ export class QueryGridDataProvider implements IGridDataProvider {
 	private async handleCopyRequestByProvider(selections: Slick.Range[], includeHeaders?: boolean): Promise<void> {
 		executeCopyWithNotification(this._notificationService, this._configurationService, selections, async () => {
 			let results = await this.queryRunner.copyResults(selections, this.batchId, this.resultSetId, this.shouldIncludeHeaders(includeHeaders));
-			let clipboardData: ClipboardData = {
-				text: results.results
+			if (results.results) {
+				let clipboardData: ClipboardData = {
+					text: results.results
+				}
+				this._clipboardService.write(clipboardData);
 			}
-			this._clipboardService.write(clipboardData);
 		});
 	}