… extensions/azuremonitor/yarn.lock
… extensions/datavirtualization/yarn.lock
… extensions/import/yarn.lock
… extensions/kusto/yarn.lock
… extensions/mssql/yarn.lock
… extensions/sql-migration/yarn.lock