<!-- Thank you for submitting a Pull Request. Please:
* Read our Pull Request guidelines:
  https://github.com/Microsoft/azuredatastudio/wiki/How-to-Contribute#pull-requests.
* Associate an issue with the Pull Request.
* Ensure that the code is up-to-date with the `main` branch.
* Include a description of the proposed changes and how to test them.
-->

This PR will fix https://github.com/microsoft/azuredatastudio/issues/25632

The other PR's needed for this PR:
sqlops-dataprotocolclient - https://github.com/microsoft/sqlops-dataprotocolclient/pull/101
sqltoolsservice - https://github.com/microsoft/sqltoolsservice/pull/2385

The reason copying was very slow from the SQL Tools Service was because `ToString` was repeatedly being called on the `StringBuilder` before `EndsWith`: https://github.com/microsoft/sqltoolsservice/blob/18ffbe6d76c511881a76f636d2e59467dfe8c4d6/src/Microsoft.SqlTools.ServiceLayer/QueryExecution/QueryExecutionService.cs#L866

In the PR for the SQL Tools Service `ToString` is only called once to correct this issue. With this correction in place, copying from STS is significantly faster, as can be seen in this gif where copying is done in the SQL Tools Service process.

In this gif 100,000 query results are being copied Through the SQL Tools Service:
![Copy 100000 query results STS](https://github.com/user-attachments/assets/8d337da9-850f-4b09-8075-9d2db88797c7)

For comparison, this is what copying 100,000 query results from SSMS looks like:
![Copy 100000 query results SSMS](https://github.com/user-attachments/assets/2dd83d04-08a3-4d29-a363-15473f95d434)

Without the changes associated with this PR and the related PR's were made, it takes a very long time to copy query results. this following gif runs for over a minute and doesn't complete the copy operation for 100,000 query results.
![Copy 100000 query results STS before change](https://github.com/user-attachments/assets/db05e4ad-4e85-4177-b673-7ca0ca615845)

## Copying on Linux

To ensure that copying works on the Linux platform and to avoid repeating this issue (https://github.com/microsoft/azuredatastudio/issues/24043), copying defaults to the UI process. This is because the backend SQL Tools Service uses a NuGet package named `TextCopy` to copy query results. When copying on Linux, `TextCopy` depends on `xsel` and not all Linux users may have that installed, and as a result will find that copying doesn't work without `xsel`:
![image](https://github.com/user-attachments/assets/9fdca510-a73f-4d1c-aa4c-0dcf3dd0f61d)


## Time Differences between copying from ADS UI and STS
### 6 Column Query Results:

>SELECT TOP [25000|50000|75000|100000] sys.all_columns.name, column_id, system_type_id, user_type_id, max_length, [precision]
FROM sys.all_columns, sys.all_views;

Copying results with 6 columns from the UI:
ADS UI Copy with 25,000 rows: 813.864013671875 ms
ADS UI Copy with 50,000 rows: 1506.052001953125 ms
ADS UI Copy with 75,000 rows: 2277.085205078125 ms
ADS UI Copy with 100,000 rows: 2966.846923828125 ms

Copying results with 6 columns from STS:
STS Copy with 25,000 rows: 0.379150390625 ms
STS Copy with 50,000 rows: 0.408935546875 ms
STS Copy with 75,000 rows: 0.454833984375 ms
STS Copy with 100,000 rows: 0.570068359375 ms



### 53 Column Query Results:
> SELECT TOP [25000|50000|75000|100000] *
FROM sys.all_columns, sys.all_views;

Copying results with 53 columns from the UI:
ADS UI Copy with 25,000 rows: 12446.373046875 ms
ADS UI Copy with 50,000 rows: 24727.3408203125 ms
ADS UI Copy with 75,000 rows: 38170.3740234375 ms
ADS UI Copy with 100,000 rows: 51372.80810546875 ms


Copying results with 53 columns from STS:
STS Copy with 25,000 rows: 0.575927734375 ms
STS Copy with 50,000 rows: 0.652099609375 ms
STS Copy with 75,000 rows: 0.81201171875 ms
STS Copy with 100,000 rows: 1.3330078125 ms