@caohai 2024-08-29
So this change (alongside with PRs in STS + sqlops-dataprotocolclient) will now do the two following things, right?
1) revert the provider side copy feature to the state Alan initially implemented.
2) set the default value for provider side copy config to `false` instead of `true`.

This means win/mac users will need to manually enable this config to get the fast copy feature. Can we enable this for win/mac automatically without breaking copy for linux?

Can we do something like this?
```
'default': isLinux ? false : true,
```