@kburtram 2024-08-29
> Win/Mac users don't need to change anything because copying from the UI was much faster than copying from the query provider.

I'm confused by this comment.  What is "query provider" referring to here, the query service in STS?  The original bug for slow copy was reported on Windows with the current implementation (https://github.com/microsoft/azuredatastudio/issues/25632). What is this PR doing to improve the copy performance on Windows?