@kburtram 2024-08-29
> > > So this change (alongside with PRs in STS + sqlops-dataprotocolclient) will now do the two following things, right?
> > > 
> > > 1. revert the provider side copy feature to the state Alan initially implemented.
> > > 2. set the default value for provider side copy config to `false` instead of `true`.
> > > 
> > > This means win/mac users will need to manually enable this config to get the fast copy feature. Can we enable this for win/mac automatically without breaking copy for linux?
> > > Can we do something like this?
> > > ```
> > > 'default': isLinux ? false : true,
> > > ```
> > 
> > 
> > This change will
> > 
> > 1. Revert the provider side copy feature to the state Alan initially implemented
> > 2. will also set the query provider copy feature to false.
> > 
> > Win/Mac users don't need to change anything because copying from the UI was much faster than copying from the query provider.
> 
> I was always under the impression that STS side copy is faster and not aware of the UI side copy optimization happened in #25152. Can you update the PR description to include the whole story?

It would also be interesting to see this PR compared to the current implementation.  The two changes I see here for default Windows users is that the copy method is now awaited and the result string is no longer returned from STS (it's not clear that's even being called in default Windows code path).  The recent bug reports are on Windows after https://github.com/microsoft/azuredatastudio/pull/25152, so unclear how this change address this bug to me.